incentives cleanup. wired rollapp creation hook. ut compiles
