+++ x/incentives/keeper/gauge_test.go
@@ -83,10 +83,8 @@ func (suite *KeeperTestSuite) TestGaugeOperations() {
 		suite.Require().Len(gaugeIdsByDenom, 0)
 
 		// setup lock and gauge
-		lockOwners := suite.SetupManyLocks(tc.numLocks, defaultLiquidTokens, defaultLPTokens, time.Second)
+		_ = suite.SetupManyLocks(tc.numLocks, defaultLiquidTokens, defaultLPTokens, time.Second)
 		gaugeID, _, coins, startTime := suite.SetupNewGauge(tc.isPerpetual, sdk.Coins{sdk.NewInt64Coin("stake", 12)})
-		// evenly distributed per lock
-		expectedCoinsPerLock := sdk.Coins{sdk.NewInt64Coin("stake", 12/int64(tc.numLocks))}
 		// set expected epochs
 		var expectedNumEpochsPaidOver int
 		if tc.isPerpetual {
@@ -118,10 +116,6 @@ func (suite *KeeperTestSuite) TestGaugeOperations() {
 		gaugeIdsByDenom = suite.App.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.Ctx, "lptoken")
 		suite.Require().Len(gaugeIdsByDenom, 1)
 		suite.Require().Equal(gaugeID, gaugeIdsByDenom[0])
-
-		// check rewards estimation
-		rewardsEst := suite.App.IncentivesKeeper.GetRewardsEst(suite.Ctx, lockOwners[0], []lockuptypes.PeriodLock{}, 100)
-		suite.Require().Equal(expectedCoinsPerLock.String(), rewardsEst.String())
 
 		// check gauges
 		gauges = suite.App.IncentivesKeeper.GetNotFinishedGauges(suite.Ctx)
@@ -215,8 +209,6 @@ func (suite *KeeperTestSuite) TestGaugeOperations() {
 			// check invalid gauge ID
 			_, err = suite.App.IncentivesKeeper.GetGaugeByID(suite.Ctx, gaugeID+1000)
 			suite.Require().Error(err)
-			rewardsEst = suite.App.IncentivesKeeper.GetRewardsEst(suite.Ctx, lockOwners[0], []lockuptypes.PeriodLock{}, 100)
-			suite.Require().Equal(sdk.Coins{}, rewardsEst)
 
 			// check gauge ids by denom
 			gaugeIdsByDenom = suite.App.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.Ctx, "lptoken")
@@ -226,10 +218,6 @@ func (suite *KeeperTestSuite) TestGaugeOperations() {
 			// check finished gauges
 			gauges = suite.App.IncentivesKeeper.GetFinishedGauges(suite.Ctx)
 			suite.Require().Len(gauges, 0)
-
-			// check rewards estimation
-			rewardsEst = suite.App.IncentivesKeeper.GetRewardsEst(suite.Ctx, lockOwners[0], []lockuptypes.PeriodLock{}, 100)
-			suite.Require().Equal(sdk.Coins(nil), rewardsEst)
 
 			// check gauge ids by denom
 			gaugeIdsByDenom = suite.App.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.Ctx, "lptoken")
